+++ ghost/core/core/server/services/api-version-compatibility/index.js
@@ -1,5 +1,5 @@
 const APIVersionCompatibilityService = require('./APIVersionCompatibilityService');
-const versionMismatchHandler = require('@tryghost/mw-api-version-mismatch');
+const versionMismatchHandler = require('./mw-api-version-mismatch');
 const ghostVersion = require('@tryghost/version');
 const {GhostMailer} = require('../mail');
 const settingsService = require('../settings/settings-service');
