+++ ghost/core/package.json
@@ -119,7 +119,6 @@
     "@tryghost/metrics": "1.0.37",
     "@tryghost/milestones": "0.0.0",
     "@tryghost/minifier": "0.0.0",
-    "@tryghost/mw-api-version-mismatch": "0.0.0",
     "@tryghost/mw-error-handler": "0.0.0",
     "@tryghost/mw-version-match": "0.0.0",
     "@tryghost/mw-vhost": "0.0.0",
