--- ghost/mw-api-version-mismatch/README.md
@@ -1,3 +1,0 @@
-# Mw Api Version Mismatch
-
-Middleware handling mismatched Accept-Version header
