--- ghost/mw-api-version-mismatch/index.js
@@ -1,1 +1,0 @@
-module.exports = require('./lib/mw-api-version-mismatch');
