--- ghost/mw-api-version-mismatch/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "@tryghost/mw-api-version-mismatch",
-  "private": true,
-  "version": "0.0.0",
-  "author": "Ghost Foundation",
-  "license": "MIT",
-  "main": "index.js",
-  "scripts": {
-    "dev": "echo \"Implement me!\"",
-    "test:unit": "NODE_ENV=testing c8 --all --check-coverage --reporter text --reporter cobertura -- mocha --reporter dot './test/**/*.test.js'",
-    "test": "yarn test:unit",
-    "lint:code": "eslint *.js lib/ --ext .js --cache",
-    "lint": "yarn lint:code && yarn lint:test",
-    "lint:test": "eslint -c test/.eslintrc.js test/ --ext .js --cache"
-  },
-  "files": [
-    "index.js",
-    "lib"
-  ],
-  "devDependencies": {
-    "@tryghost/errors": "1.3.5",
-    "c8": "8.0.1",
-    "mocha": "10.2.0",
-    "sinon": "15.2.0"
-  },
-  "dependencies": {
-    "jsonwebtoken": "8.5.1"
-  }
-}
